refactor: split out some components
